[refactor] App을 컴포넌트로 분리
