+++ package.json
@@ -19,12 +19,12 @@
     "@codeceptjs/configure": "^0.6.2",
     "babel-loader": "^8.2.2",
     "codeceptjs": "^3.0.6",
-    "eslint": "^7.25.0",
-    "eslint-config-airbnb": "^18.2.1",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.23.2",
-    "eslint-plugin-react-hooks": "^4.2.0",
+    "eslint": "^8.4.1",
+    "eslint-config-airbnb": "^19.0.2",
+    "eslint-plugin-import": "^2.25.3",
+    "eslint-plugin-jsx-a11y": "^6.5.1",
+    "eslint-plugin-react": "^7.27.1",
+    "eslint-plugin-react-hooks": "^4.3.0",
     "puppeteer": "^9.0.0",
     "start-server-and-test": "^1.12.1",
     "webpack": "^5.36.2",
